requirements.txt
torch>=2.0.1
torchaudio>=2.0.1
git+https://github.com/facebookresearch/audiocraft.git#egg=audiocraft
openai>=0.27.0
python-dotenv>=0.21.0
streamlit>=1.20.0
tqdm>=4.64.0